codegen: Add `-nostartfiles` to drop the hardcoded prelude
This is useful mainly if you desire to use your own "prelude" for
whatever purpose you may have.

This includes, but is not limited to, prepending a prelude that is aware
of the [metadata conventions](https://wiki.xxiivv.com/site/metadata.html).

Note that *at this point in time* it is not possible to trivially place
a prelude within the C code. The file always starts with `bss`, then `text`,
making things awkward at best.

It would be expected that the prelude is "bring your own" outside of
what this compiler produces.